sbt release
