+++ .github/workflows/release_deploy.yml
@@ -93,7 +93,7 @@ jobs:
     steps:
       - name: Make Release
         id: release
-        uses: pagopa/github-actions-template/sbt-release@main
+        uses: pagopa/github-actions-template/sbt-release@v1.11.0
         with:
           semver: ${{ needs.setup.outputs.semver }}
           github_token: ${{ secrets.BOT_TOKEN_GITHUB }}
